SS-8827: Viewport buttons refactoring
