+++ hooks/ui/useCustomTheme.ts
@@ -1089,18 +1089,31 @@ export const useCustomTheme = (props: Props = {}) => {
 			 * Which viewport icons to display.
 			 */
 			ViewportIcons: ViewportIconsThemeProps({
+				// style: {
+				// 	display: "flex",
+				// 	gap: "0.25rem",
+				// 	alignItems: "center",
+				// 	flexDirection: "row",
+				// },
+				// offset: "0.5em",
+				// py: 1,
+				// px: 0,
+				// shadow: "md",
+				// size: 32,
 				// color: "black",
 				// colorDisabled: "grey",
-				// enableArBtn: false,
-				// enableCamerasBtn: false,
-				// enableFullscreenBtn: false,
-				// enableZoomBtn: false,
-				// fullscreenId: "viewer-fullscreen-area",
-				// iconStyle: { m: "3px" },
-				// size: 32,
-				// style: { display: "flex"},
 				// variant: "subtle",
 				// variantDisabled: "transparent",
+				// iconStyle: { m: "0.188rem" },
+				// fullscreenId: "viewer-fullscreen-area",
+				// enableHistoryButtons: true,
+				// enableModelStateButtons: true,
+				// enableImportExportButtons: true,
+				// enableResetButton: true,
+				// enableArBtn: true,
+				// enableCamerasBtn: true,
+				// enableFullscreenBtn: true,
+				// enableZoomBtn: true,
 			}),
 			/**
 			 * ViewportOverlayWrapper
